requirements.txt
# Core dependencies
Django>=5.1.5
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
djangorestframework>=3.14.0

# Database adapters
dj-database-url>=2.1.0  # For database URL configuration
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Storage backends
django-storages>=1.14.2  # For S3 and other storage backends
boto3>=1.34.0  # AWS SDK for Python

# Caching
django-redis>=5.4.0  # Redis cache backend
redis>=5.0.1  # Redis client
hiredis>=2.2.3  # C parser for Redis

# Development tools
django-debug-toolbar>=4.2.0  # Debug toolbar for development

# Testing
pytest>=7.4.0  # Testing framework
pytest-django>=4.7.0  # Django plugin for pytest

# Code quality
black>=23.11.0  # Code formatter
isort>=5.12.0  # Import sorter
flake8>=6.1.0  # Linter